llm_requirements.txt
pymongo==4.6.2
llama-index
llama-index-embeddings-mistralai
llama-index-vector-stores-mongodb
llama-index-llms-mistralai
transformers==4.38.2
instructor
openai
python-dotenv
twikit
flask
